requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
celery==5.3.4
redis==5.0.1
psycopg
python-decouple==3.8
Pillow==12.0.0
stripe==7.10.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
